implementing new schema for graph serialization and getting most of the code in place for a new unsafe graph that should load + perform faster

+++ ScriptureGraph.Core/Graph/KnowledgeGraphNode.cs
@@ -8,7 +8,7 @@ namespace ScriptureGraph.Core.Graph
 {
     public class KnowledgeGraphNode
     {
-        public KnowledgeGraphNode(int edgeCapacity)
+        public KnowledgeGraphNode(ushort edgeCapacity)
         {
             Edges = new GraphEdgeList(edgeCapacity);
         }
